adding the content component to conditionally render different content
