+++ app/src/App.js
@@ -4,14 +4,13 @@ import {Nav} from './components/Nav'
 import {Posts} from './pages/Posts'
 import {Routes, Route} from 'react-router-dom'
 import styled from 'styled-components'
-import {Post} from './components/Post'
 import Header from './components/Header'
 import {RxHamburgerMenu} from 'react-icons/rx'
 import {useState} from 'react'
 import {FaAws} from 'react-icons/fa'
 import {SiKubernetes, SiMicrosoftazure, SiLinux} from 'react-icons/si'
 import {Projects} from './components/Projects'
-import {Project} from './pages/Project'
+import {Content} from './components/Content'
 
 
 function App() {
@@ -30,9 +29,9 @@ function App() {
       <Routes>
           <Route path="/" element={<Home />}/>
           <Route path="/posts" element={ <Posts />} />
-          <Route path="/posts/:id" exact={true} element={ <Post/>}/>
           <Route path="/projects" exact={true} element={ <Projects />} />
-          <Route path="/projects/:id" exact={true} element={ <Project  />} />
+          <Route path="/projects/:id" exact={true} element={ <Content/>} />
+          <Route path="/posts/:id" exact={true} element={ <Content/>}/>
       </Routes>
       </div>
       <footer>
